% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{process}
\alias{process}
\title{The process of geting the differential expression genes}
\usage{
process(dataset)
}
\arguments{
\item{dateset}{The dataset of cancer or disease from NCBI, such as GSE1145}
}
\value{
There are five variable values, exprs stand for gene expression profile,
normal stand for the number of normal samples, tumor stand for the number of tumor
samples, DE stand for the fold change of differential expression genes, ALL stand for
all genes of human
}
\description{
This function uses based method such as limma to get the differential expression gene
}
\details{
the previous SPIA method and integrate the change of
of genes Pearson coefficient(PCC) from two groups. We proposed a set of three
pathway analysis methods based on the change of PCC. We applied these approaches
to colorectal cancer, lung cancer and Alzheimer's disease datasets and so on.
}
\examples{
#import EnrichmentBrowser, KEGGandMetacoreDzPathwaysGEO, KEGGdzPathwaysGEO and SPIA package
library(EnrichmentBrowser)
library(KEGGandMetacoreDzPathwaysGEO)
library(KEGGdzPathwaysGEO)
library(SPIA)
data("GSE1145")
result <- process(GSE1145)
}
